+++ src/endpoints/common_endpoint_registry.cpp
@@ -263,7 +263,8 @@ namespace ccf
       "/code", HTTP_GET, json_read_only_adapter(get_code), no_auth_required)
       .set_auto_schema<void, GetCode::Out>()
       .set_openapi_summary(
-        "This endpoint is deprecated. Please use /gov/kv/* endpoints instead.")
+        "This endpoint is deprecated. It is replaced /gov/kv/nodes/code_ids "
+        "endpoint")
       .set_openapi_deprecated(true)
       .set_openapi_description("Permitted SGX code identities")
       .install();
